# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Portal Manajemen Proyek API
  version: 1.0.0
  description: REST API untuk autentikasi, RBAC, CRUD proyek, tugas, dashboard, risiko, time tracking, file, dan komentar.
servers:
  - url: http://localhost:4000/api
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /teams/{id}:
    put:
      tags: [Dashboard]
      summary: Mengubah tim pada proyek yang dikelola Project Manager
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Tim diperbarui
    delete:
      tags: [Dashboard]
      summary: Menghapus tim pada proyek yang dikelola Project Manager
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Tim dihapus
...
  /time-logs/{id}:
    put:
      tags: [Dashboard]
      summary: Mengubah log waktu, Project Manager atau pemilik log Developer
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Log waktu diperbarui
    delete:
      tags: [Dashboard]
      summary: Menghapus log waktu, Project Manager atau pemilik log Developer
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Log waktu dihapus
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: integer
...